% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histWithDist_function.R
\name{histWithDist}
\alias{histWithDist}
\title{histWithDist}
\usage{
histWithDist(x, main = "Histogram with fitted distributions",
  distr = c("nbinom", "pois", "norm", "lnorm"))
}
\arguments{
\item{x}{Numeric vector of data values to be plotted to be plotted as a histogram.}

\item{main}{Title for the plot. Default value is \code{"Histogram with fitted distributions"}.}

\item{distr}{Character vector of distribution curves to overlay on the histogram. Note that this uses the standard R names for the distributions which differ from those used in emon. Should be one or more of "norm" (normal), "pois" (Poisson), "lnorm" (log normal) and "nbinom" (negative binomial). By default all four are plotted.}
}
\value{
Primarily used directly plotting, but also invisibly returns a \code{histWithDist} object that can be stored for later plotting if needed.
}
\description{
Plots histogram with fitted distribution curves
}
\details{
This function uses \code{MASS::fitdistr} to fit distribution curves to \code{x} for the distributions specified, then overlays these on a histogram of \code{x}.
}
\author{
James Hutchison
}

