% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvasivesCheck_function.R
\name{invasivesCheck}
\alias{invasivesCheck}
\title{invasivesCheck}
\usage{
invasivesCheck(invasives, file, worksheet)
}
\arguments{
\item{invasives}{A vector of invasive species binomial names.}

\item{file}{Filepath to the .xlsx spreadsheet file to be checked for invasive species.}

\item{worksheet}{Number or name of the worksheet in the excel file to be checked.}
}
\description{
invasivesCheck
}
\details{
This function searches for all of the values of a character vector in a worksheet of an Excel file. The intended use is for checking for invasive species in spreadsheets received from contractors. The invasive species should be supplied as a vector of species names, and the data to be searched should be in an Excel file.
}
\author{
James Hutchison
}

