% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.histWithDist_function.R
\name{plot.histWithDist}
\alias{plot.histWithDist}
\title{plot.histWithDist}
\usage{
\method{plot}{histWithDist}(x)
}
\arguments{
\item{x}{\code{histWithDist} object}
}
\description{
Plotting method for \code{histWithDist} object class
}
\details{
Plots histogram with distribution curves and legend from \code{histWithDist} object.
}
\author{
James Hutchison
}

