% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multiQQ_function.R
\name{plot.multiQQ}
\alias{plot.multiQQ}
\title{plot.multiQQ}
\usage{
\method{plot}{multiQQ}(x, y, ...)
}
\arguments{
\item{x}{\code{multiQQ} object for plotting.}
}
\description{
Plotting method for multiQQ object
}
\details{
Generates quantile-quantile plot with 1:1 line from \code{multiQQ} object.
}
\author{
James Hutchison
}

