% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.powerByFactor_function.R
\name{plot.powerByFactor}
\alias{plot.powerByFactor}
\title{plot.powerByFactor}
\usage{
\method{plot}{powerByFactor}(x, y, ...)
}
\arguments{
\item{x}{\code{powerByFactor} object}
}
\description{
Plotting method for \code{powerByFactor} object class
}
\details{
Plots graphs of statistical power (i.e. proportion of significant results) against number of samples from a \code{powerByFactor} object.
}

