% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotByFactor_function.r
\name{plotByFactor}
\alias{plotByFactor}
\title{plotByFactor}
\usage{
plotByFactor(x, f, plotFunction)
}
\arguments{
\item{x}{numeric vector of data to be plotted}

\item{f}{factor vector of same length as \code{x}, which is used to subset the data}

\item{plotFunction}{Function used to plot the data.}
}
\description{
Splits a dataset by a factor then runs a plotting function on each subset, plotting the results in a grid.
}
\author{
James Hutchison
}

