% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.groupsByFactor_function.r
\name{power.groupsByFactor}
\alias{power.groupsByFactor}
\title{power.groupsByFactor}
\usage{
power.groupsByFactor(x, f, change, change.type, n1, n2, distribution, test,
  alternative = "two.sided", alpha = 0.05, nsims = 1000, nreps = 999)
}
\arguments{
\item{x}{Numerical vector from which the distribution parameters will be calculated using \code{fitdistr}}

\item{f}{Factor the same length as \code{x}, which is used to subset the data}

\item{change}{Mean of second group minus mean of first group (i.e. mu2-mu1) or percentage change in mu1 to create mu2 (depending on value of \code{change.type}).}

\item{change.type}{Whether the parameter change represents an additive (\code{"A"}) or percentage (\code{"M"}) change.}

\item{n1}{Vector of sample sizes for group 1. Must be of same dimension as \code{n2}.}

\item{n2}{Vector of sample sizes for group 2. Must be of same dimension as \code{n1}.}

\item{distribution}{The statistical distribution for the two groups. Can be either: \code{"Normal"}, \code{"Poisson"}, \code{"Lognormal"} or \code{"Negbin"}.}

\item{test}{Should be \code{"P"} for parametric or \code{"NP"} for non-parametric. See \code{?power.groups} for more details.}

\item{alternative}{A character string specifying the alternative hypothesis, must be one of \code{"two.sided"} (default), \code{"greater"} or \code{"less"}. You can specify just the initial letter. As an example, "less" means that the alternative is that the mean of the first group is less than the mean of the second.}

\item{alpha}{The type 1 error for assessing statistical significance (default is 0.05) in the power simulations.}

\item{nsims}{Number of repeat simulations to estimate power (default is 1000).}

\item{nreps}{Number of repeat permutations for randomisation test (default is 999).}
}
\value{
An object of class \code{powerByFactor}, which is essentially a list containing the power results for each level of \code{f}. The package includes methods to \code{plot() powerByFactor} objects.
}
\description{
Subsets a dataset by a factor then uses power.groups to calculate statistical power for each group
}
\details{
This function splits \code{x} into subsets based on \code{f}, then runs \code{power.groups} on each subset. \code{pars1} and \code{pars2} for \code{power.groups} are automatically calculated based on \code{x} and \code{distribution} using basic summary stats (e.g. \code{mean}, \code{sd}) or using \code{MASS::fitdistr} if \code{distribution = "Negbin"}.
}
\author{
James Hutchison
}

