% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultibeamToPolygon_function.R
\name{multibeamToPolygon}
\alias{multibeamToPolygon}
\title{Multibeam to polygon conversion}
\usage{
multibeamToPolygon(multibeam = NULL, multibeamType = NA, polyfile = NULL,
  polyformat = "ESRI Shapefile")
}
\arguments{
\item{multibeam}{Filepath to multibeam raster file. See gdalDrivers() for a list of readable formats. Default is NULL and function will then prompt for the file location using choose.files().}

\item{multibeamType}{Raster type - should be 1 for RGB multi-band file, or 2 for floating point single band. Default is NULL and function will prompt for it using readline.}

\item{polyfile}{Filepath and name for output file location. Default is NULL and function will then prompt for it using choose.dir() and readline().}

\item{polyformat}{File type for output polygon file. Default is ESRI Shapefile. See ogrDrivers() for a list of available formats.}
}
\description{
Converts a raster of multibeam data to a single dissolved polygons showing the outline of the data
}
\author{
James Hutchison
}

